+++ app/src/main/java/com/techcourse/support/web/mapping/ManualHandlerMapping.java
@@ -37,7 +37,7 @@ public class ManualHandlerMapping implements HandlerMapping {
 
     @Override
     public Object getHandler(final HttpServletRequest request) {
-        log.debug("Request Mapping Uri : {}", request);
+        log.debug("Request Mapping Uri : {}", request.getRequestURI());
         return controllers.get(request.getRequestURI());
     }
 }
