+++ mvc/src/main/java/webmvc/org/springframework/web/servlet/mvc/tobe/ManualHandlerAdapter.java
@@ -1,20 +1,19 @@
-package com.techcourse.support.web.adapter;
+package webmvc.org.springframework.web.servlet.mvc.tobe;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import webmvc.org.springframework.web.servlet.ModelAndView;
 import webmvc.org.springframework.web.servlet.mvc.asis.Controller;
-import webmvc.org.springframework.web.servlet.mvc.tobe.HandlerAdapter;
 import webmvc.org.springframework.web.servlet.view.JspView;
 
-public class ManualHanlderAdapter implements HandlerAdapter {
+public class ManualHandlerAdapter implements HandlerAdapter {
     @Override
     public boolean supports(final Object handler) {
         return handler instanceof Controller;
     }
 
     @Override
-    public ModelAndView handle(final Object handler, final HttpServletRequest request, final HttpServletResponse response) throws Exception{
+    public ModelAndView handle(final Object handler, final HttpServletRequest request, final HttpServletResponse response) throws Exception {
         final Controller controller = (Controller) handler;
         final String path = controller.execute(request, response);
         return new ModelAndView(new JspView(path));
